+++ src/main/java/ru/nf/conveyor/model/LoanOfferDTO.java
@@ -17,12 +17,12 @@ public class LoanOfferDTO {
 	/**
 	 * Запрашиваемая сумма
 	 */
-	private BigDecimal requestedAmount;
+	private Double requestedAmount;
 
 	/**
 	 * Общая сумма
 	 */
-	private BigDecimal totalAmount;
+	private Double totalAmount;
 
 	/**
 	 * Срок кредита
@@ -32,12 +32,12 @@ public class LoanOfferDTO {
 	/**
 	 * Ежемесячный платеж
 	 */
-	private BigDecimal monthlyPayment;
+	private Double monthlyPayment;
 
 	/**
 	 * Ставка
 	 */
-	private BigDecimal rate;
+	private Double rate;
 
 	/**
 	 * Включена страховка?
